+++ src/Infrastructure/Migrations/20200301091030_FixedIntialize.cs
@@ -1,0 +1,27 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace Infrastructure.Migrations
+{
+    public partial class FixedIntialize : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<decimal>(
+                name: "Price",
+                table: "Hotel",
+                type: "decimal(6,1)",
+                nullable: false,
+                oldClrType: typeof(decimal));
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<decimal>(
+                name: "Price",
+                table: "Hotel",
+                nullable: false,
+                oldClrType: typeof(decimal),
+                oldType: "decimal(6,1)");
+        }
+    }
+}