+++ changelogs/CHANGELOG.md
@@ -28,3 +28,23 @@ Configuration files fixed. Renaming due to convention.
 
 - ValueGeneratedOnAdd removed.
 
+# 2020-03-1
+
+Migrations added and sql server conecting handled by @bornaamir8@gmail.com .
+
+# Unrealeased
+
+-Apply Configurations dynamiclly.
+
+# Added
+- Migration files 
+- Apply  all configuration in packageDbcontex staticlly by @bornaamir8@gmail.com.
+- Connection string in appsetting.json  and  it's configuration in startup.cs by @bornaamir8@gmail.com
+- Dbcontext option builder methods in statrup.cs in configuresqlserver method by @bornaamir8@gmail.com
+- EntityFramework Directory in Infrastructure by @bornaamir8@gmail.com
+
+# Changed
+- Base configuration by @bornaamir8@gmail.com
+- Hotel configuration by @bornaamir8@gmail.com
+
+